% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_random}
\alias{create_random}
\title{Creates a random BEAST2 setup to validate}
\usage{

  create_random(input_fasta_filename = beautier::get_beautier_path("anthus_aco.fas"))
}
\description{
Creates a random BEAST2 setup to validate
}
