% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rnd_anything.R
\name{create_rnd_anything}
\alias{create_rnd_anything}
\title{Creates something}
\usage{
create_rnd_anything(
  fasta_filename = beautier::get_beautier_path("anthus_aco.fas")
)
}
\description{
Creates something
}
