% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rnd_clock_model_nasty.R
\name{create_rnd_clock_model_nasty}
\alias{create_rnd_clock_model_nasty}
\title{May create a random clock model}
\usage{
create_rnd_clock_model_nasty()
}
\description{
May create a random clock model
}
