% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_rnd_inference_model}
\alias{create_rnd_inference_model}
\title{Create a random inference model}
\usage{
create_rnd_inference_model(fasta_filename)
}
\arguments{
\item{fasta_filename}{a FASTA filename}
}
\description{
Create a random inference model
}
