% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_rnd_mcmc_chain_length}
\alias{create_rnd_mcmc_chain_length}
\title{Create a random MCMC chain length}
\usage{
create_rnd_mcmc_chain_length()
}
\description{
Create a random MCMC chain length
}
