% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_rnd_mcmc_store_every}
\alias{create_rnd_mcmc_store_every}
\title{Create a random MCMC store interval}
\usage{
create_rnd_mcmc_store_every()
}
\description{
Create a random MCMC store interval
}
