% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rnd_mrca_prior_nasty.R
\name{create_rnd_mrca_prior_nasty}
\alias{create_rnd_mrca_prior_nasty}
\title{May create a random MRCA prior}
\usage{
create_rnd_mrca_prior_nasty(
  fasta_filename = beautier::get_beautier_path("anthus_aco.fas")
)
}
\arguments{
\item{fasta_filename}{a FASTA filename}
}
\description{
May create a random MRCA prior
}
