% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rnd_tree_prior_nasty.R
\name{create_rnd_tree_prior_nasty}
\alias{create_rnd_tree_prior_nasty}
\title{May create a random tree prior}
\usage{
create_rnd_tree_prior_nasty()
}
\description{
May create a random tree prior
}
